Created a way to store a bitmap of data about blocks.
BlockBitmap is a way to keep a sparse map of booleans, in a single
file, broken up into sparse segments of 16x16x16 blocks.

Each sparse segment represents 4096 blocks, and takes up 512 bytes
(plus a 24 byte header making 524 bytes), thus striking a compromise in
time and space storage for this kind of data.

In real usage, this data will be clustered naturally as people build,
so it's not a big deal that each segment is 512 bytes.