Add in WorldData system to allow separating world-scoped data
There are many coordinate-based data structures, and to add in a world
lookup to every single one of them is unnecessarily complex. For many
servers, there's only the default 3 worlds, and even for servers with
lots of multiword, this system saves a lot of extra lookups and reduces
set cardinality.

This is done by creating a WorldData class which manages a folder for a
given world, based on an MD5 encoding of the world name, and also
manages any structures below it as needed. WorldData instances are
created on demand the first time data for a given world is requested.